+++ Common/Figures/Rectangle.hpp
@@ -6,6 +6,6 @@ struct Rectangle : Shape
 {
     std::string toJson() const override;
 
-    double sizeA = 1;
-    double sizeB = 1;
+    int sizeA = 1;
+    int sizeB = 1;
 };
